+++ testng-eclipse-plugin/src/main/org/testng/eclipse/TestNGPluginConstants.java
@@ -14,16 +14,12 @@ public abstract class TestNGPluginConstants {
   public static final String MAIN_RUNNER = RemoteTestNG.class.getName();
   
   public static final String S_OUTDIR= ".outdir"; //$NON-NLS-1$
-  public static final String S_ABSOLUTEPATH= ".absolutepath"; //$NON-NLS-1$
 //  public static final String S_REPORTERS = ".reporters";
   public static final String S_DISABLEDLISTENERS = ".disabledListeners";
 //  public static final String S_PARALLEL = ".parallel";
   public static final String S_WATCH_RESULTS = ".watchResults";
   public static final String S_WATCH_RESULT_DIRECTORY = ".watchResultDirectory";
   
-  public static final String S_DEPRECATED_OUTPUT = "generalOutput";
-  public static final String S_DEPRECATED_ABSOLUTEPATH = "generalOutputRelative";
-
   public static final String S_FAILED_TESTS = "failedTests";
   public static final String S_XML_TEMPLATE_FILE = "xmlTemplateFile";
   public static final String S_JVM_ARGS = "jvmArgs";